+++ src/classes/features/Timer.ts
@@ -1,15 +1,14 @@
 import { ModCallback } from "isaac-typescript-definitions";
-import { Callback } from "isaacscript-common";
-import { isTimerEnabled } from "../../config";
+import { Callback, ModFeature } from "isaacscript-common";
 import { getModifiedBossID } from "../../enums/BossIDCustom";
 import { TimerType } from "../../enums/TimerType";
 import { timerDraw } from "../../timer";
 import { getNumSecondsForBossObjective } from "../../types/Objective";
-import { RandomizerModFeature } from "../RandomizerModFeature";
 import { getSecondsSinceLastDamage } from "./BossNoHitObjectiveDetection";
 import { getPlaythroughSecondsElapsed } from "./StatsTracker";
 
-export class Timer extends RandomizerModFeature {
+// TODO: RANDOMIZER MOD FEATURE
+export class Timer extends ModFeature {
   @Callback(ModCallback.POST_RENDER)
   postRender(): void {
     this.drawMainTimer();
@@ -17,10 +16,10 @@ export class Timer extends RandomizerModFeature {
   }
 
   drawMainTimer(): void {
-    if (isTimerEnabled()) {
-      const seconds = getPlaythroughSecondsElapsed();
-      timerDraw(TimerType.MAIN, seconds);
-    }
+    // if (isTimerEnabled()) {
+    const seconds = getPlaythroughSecondsElapsed();
+    timerDraw(TimerType.MAIN, seconds);
+    // }
   }
 
   drawNoHitTimer(): void {