+++ docs/history.md
@@ -13,6 +13,7 @@
 - All enemies now have a chance to become a champion.
 - Added the "randomizerVersion" command.
 - Black Markets will now unlock before the first Devil Room pool collectible.
+- Fixed the bug where the timer would not work properly after 10 hours. (Thanks Neonomi)
 
 ### _v0.6.0_ - October 20th, 2023